% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{as_matrix}
\alias{as_matrix}
\title{For converting a sparse matrix into a dense matrix. Use this when after obtaining log normalised count data from either singlecelldata objects or Seurat objects}
\usage{
as_matrix(mat)
}
\arguments{
\item{mat}{a sparse matrix that is to be converted to a dense matrix}
}
\description{
For converting a sparse matrix into a dense matrix. Use this when after obtaining log normalised count data from either singlecelldata objects or Seurat objects
}
