% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{NMAD}
\alias{NMAD}
\title{NMAD is for calculating normalised MAD, where the MAD value is divided by the mean. This allows for comparison of NMAD results between different datasets.}
\usage{
NMAD(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
NMAD is for calculating normalised MAD, where the MAD value is divided by the mean. This allows for comparison of NMAD results between different datasets.
}
