% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{RMOutlier}
\alias{RMOutlier}
\title{RMOutlier is for the removal of outliers within gene expression datasets. Users can alter the cutoffs for what is deemed to be an outlier using the STD variable. The STD variable denotes how many standard deviations from teh median.}
\usage{
RMOutlier(x, STD = 4)
}
\arguments{
\item{x}{Vector of values}

\item{STD}{number of deviations from median. Default set at 4}
}
\description{
RMOutlier is for the removal of outliers within gene expression datasets. Users can alter the cutoffs for what is deemed to be an outlier using the STD variable. The STD variable denotes how many standard deviations from teh median.
}
