% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{scDS.Pval}
\alias{scDS.Pval}
\title{scDS.Pval an method for scQuant to calculate genes that are differentially stable (DS) between cell types. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "DS.Pval")}
\usage{
scDS.Pval(dat.list = dat.list)
}
\arguments{
\item{dat.list}{Output from scQuant (method = "std")}
}
\description{
scDS.Pval an method for scQuant to calculate genes that are differentially stable (DS) between cell types. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "DS.Pval")
}
