% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{scDS}
\alias{scDS}
\title{scDS.Pval an method for scQuant to calculate genes that are differentially stable (DS) between cell types. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "DS")}
\usage{
scDS(dat.list, QuantNo)
}
\arguments{
\item{dat.list}{Output from scQuant (method = "std")}

\item{QuantNo}{Number of quantiles that were used to generate dat.list}
}
\description{
scDS.Pval an method for scQuant to calculate genes that are differentially stable (DS) between cell types. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "DS")
}
