% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{scQuant}
\alias{scQuant}
\title{Computing quantile normalised MAD}
\usage{
scQuant(
  geneMat = geneMat,
  celltype = celltype,
  zeroPerc = zeroPerc,
  OutlierRemoval = TRUE,
  QuantNo = QuantNo,
  method = c("SG", "SV", "DS", "std", "NoPval")
)
}
\arguments{
\item{geneMat}{Expression matrix where columns are cells and rows are genes}

\item{celltype}{Vector of cell type labels, must correspond to columns in geneMat}

\item{zeroPerc}{Percentage of lowly expressed cells that need to be removed}

\item{OutlierRemoval}{Removal of outlier gene expression set at 4x of MAD values
Default set to TRUE}

\item{QuantNo}{Total number of quantiles the gene expression distribution is to be spli into}

\item{method}{Different analysis building upon the quantile methodology. "std", returns the quantile NMAD values for each gene, for each cell type and a list of data frames. "SG" calculates genes with and ranks genes with based on how stable the gene expressions are, genes are ranked from most to least stable, where there higher the ranking the more stable the genes are. "DS" calculates genes that are differentially stably expressed between cell types, using two-sample Wilcoxon test. "NoPval", calculates differentially stable genes by comparing the ranking between genes expression from different cell types."SV" calculate and returns a ranking of genes that are stably variable either between cell types or between datasets}
}
\value{
with std returns the quantile NMAD values
}
\description{
Computing quantile normalised MAD
}
