% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{scSG}
\alias{scSG}
\title{scSG an method for scQuant to calculate and rank stable genes. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "std")}
\usage{
scSG(dat.list = dat.list, QuantNo = QuantNo)
}
\arguments{
\item{dat.list}{Output from scQuant (method = "std")}

\item{QuantNo}{Number of quantiles that were used to generate dat.list}
}
\value{
ranking of stably expressed genes across different cell types
}
\description{
scSG an method for scQuant to calculate and rank stable genes. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "std")
}
