% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQuant.R
\name{scSV}
\alias{scSV}
\title{scSV an method for scQuant to calculate and rank genes that are similarly variable. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "SV")}
\usage{
scSV(dat.list = dat.list, QuantNo = QuantNo)
}
\arguments{
\item{dat.list}{Output from scQuant (method = "std")}

\item{QuantNo}{Number of quantiles that were used to generate dat.list}
}
\description{
scSV an method for scQuant to calculate and rank genes that are similarly variable. This function can be run within scQuant or be used as a standalone, to be run using the results from scQuant (method = "SV")
}
